Rename grpc stream classes
