+++ src/ApiCallable.php
@@ -146,19 +146,19 @@ class ApiCallable
             case 'ClientStreaming':
                 $apiCall = function () use ($callable, $grpcStreamingDescriptor) {
                     $response = ApiCallable::callWithoutRequest($callable, func_get_args());
-                    return new ClientStreamingResponse($response, $grpcStreamingDescriptor);
+                    return new ClientStream($response, $grpcStreamingDescriptor);
                 };
                 break;
             case 'ServerStreaming':
                 $apiCall = function () use ($callable, $grpcStreamingDescriptor) {
                     $response = call_user_func_array($callable, func_get_args());
-                    return new ServerStreamingResponse($response, $grpcStreamingDescriptor);
+                    return new ServerStream($response, $grpcStreamingDescriptor);
                 };
                 break;
             case 'BidiStreaming':
                 $apiCall = function () use ($callable, $grpcStreamingDescriptor) {
                     $response = ApiCallable::callWithoutRequest($callable, func_get_args());
-                    return new BidiStreamingResponse($response, $grpcStreamingDescriptor);
+                    return new BidiStream($response, $grpcStreamingDescriptor);
                 };
                 break;
             default:
